componentizing metadata table-tree
